Read tree from release-4-8-4

+++ ipaserver/install/dsinstance.py
@@ -24,12 +24,15 @@ import logging
 import shutil
 import pwd
 import os
-import re
 import time
 import tempfile
 import fnmatch
 
-import ldap
+from lib389 import DirSrv
+from lib389.idm.ipadomain import IpaDomain
+from lib389.instance.options import General2Base, Slapd2Base
+from lib389.instance.remove import remove_ds_instance as lib389_remove_ds
+from lib389.instance.setup import SetupDs
 
 from ipalib import x509
 from ipalib.install import certmonger, certstore
@@ -95,27 +98,22 @@ def schema_dirname(serverid):
     return config_dirname(serverid) + "/schema/"
 
 
-def remove_ds_instance(serverid, force=False):
-    """A wrapper around the 'remove-ds.pl' script used by
-    389ds to remove a single directory server instance. In case of error
-    additional call with the '-f' flag is performed (forced removal). If this
-    also fails, then an exception is raised.
+def remove_ds_instance(serverid):
+    """Call the lib389 api to remove the instance. Because of the
+    design of the api, there is no "force" command. Provided a marker
+    file exists, it will attempt the removal, and the marker is the *last*
+    file to be removed. IE just run this multiple times til it works (if
+    you even need multiple times ....)
     """
-    instance_name = ''.join([DS_INSTANCE_PREFIX, serverid])
-    args = [paths.REMOVE_DS_PL, '-i', instance_name]
-    if force:
-        args.append('-f')
-        logger.debug("Forcing instance removal")
 
-    try:
-        ipautil.run(args)
-    except ipautil.CalledProcessError:
-        if force:
-            logger.error("Instance removal failed.")
-            raise
-        logger.debug("'%s' failed. "
-                     "Attempting to force removal", paths.REMOVE_DS_PL)
-        remove_ds_instance(serverid, force=True)
+    logger.debug("Attempting to remove instance %s", serverid)
+    # Alloc the local instance by name (no creds needed!)
+    ds = DirSrv(verbose=True, external_log=logger)
+    ds.local_simple_allocate(serverid)
+
+    # Remove it
+    lib389_remove_ds(ds)
+    logger.debug("Instance removed correctly.")
 
 
 def get_ds_instances():
@@ -163,18 +161,17 @@ def is_ds_running(server_id=''):
 
 
 def get_domain_level(api=api):
-    ldap_uri = ipaldap.get_ldap_uri(protocol='ldapi', realm=api.env.realm)
-    conn = ipaldap.LDAPClient(ldap_uri)
-    conn.external_bind()
-
     dn = DN(('cn', 'Domain Level'),
             ('cn', 'ipa'), ('cn', 'etc'), api.env.basedn)
 
-    try:
-        entry = conn.get_entry(dn, ['ipaDomainLevel'])
-    except errors.NotFound:
-        return constants.DOMAIN_LEVEL_0
-    return int(entry.single_value['ipaDomainLevel'])
+    with ipaldap.LDAPClient.from_realm(api.env.realm) as conn:
+        conn.external_bind()
+        try:
+            entry = conn.get_entry(dn, ['ipaDomainLevel'])
+        except errors.NotFound:
+            return constants.DOMAIN_LEVEL_0
+        else:
+            return int(entry.single_value['ipaDomainLevel'])
 
 
 def get_all_external_schema_files(root):
@@ -185,32 +182,6 @@ def get_all_external_schema_files(root):
             if fnmatch.fnmatch(name, "*.ldif"):
                 f.append(os.path.join(path, name))
     return sorted(f)
-
-
-INF_TEMPLATE = """
-[General]
-FullMachineName=   $FQDN
-SuiteSpotUserID=   $USER
-SuiteSpotGroup=    $GROUP
-ServerRoot=    $SERVER_ROOT
-[slapd]
-ServerPort=   389
-ServerIdentifier=   $SERVERID
-Suffix=   $SUFFIX
-RootDN=   cn=Directory Manager
-RootDNPwd= $PASSWORD
-InstallLdifFile= /var/lib/dirsrv/boot.ldif
-inst_dir=   /var/lib/dirsrv/scripts-$SERVERID
-"""
-
-BASE_TEMPLATE = """
-dn: $SUFFIX
-objectClass: top
-objectClass: domain
-objectClass: pilotObject
-dc: $BASEDC
-info: IPA V2.0
-"""
 
 
 class DsInstance(service.Service):
@@ -243,7 +214,7 @@ class DsInstance(service.Service):
         self.domainlevel = domainlevel
         if realm_name:
             self.suffix = ipautil.realm_to_suffix(self.realm)
-            self.serverid = installutils.realm_to_serverid(self.realm)
+            self.serverid = ipaldap.realm_to_serverid(self.realm)
             self.__setup_sub_dict()
         else:
             self.suffix = DN()
@@ -254,7 +225,6 @@ class DsInstance(service.Service):
     def __common_setup(self):
 
         self.step("creating directory server instance", self.__create_instance)
-        self.step("enabling ldapi", self.__enable_ldapi)
         self.step("configure autobind for root", self.__root_autobind)
         self.step("stopping directory server", self.__stop_instance)
         self.step("updating configuration in dse.ldif", self.__update_dse_ldif)
@@ -298,7 +268,7 @@ class DsInstance(service.Service):
                   idstart, idmax, pkcs12_info, ca_file=None,
                   setup_pkinit=False):
         self.realm = realm_name.upper()
-        self.serverid = installutils.realm_to_serverid(self.realm)
+        self.serverid = ipaldap.realm_to_serverid(self.realm)
         self.suffix = ipautil.realm_to_suffix(self.realm)
         self.fqdn = fqdn
         self.dm_password = dm_password
@@ -422,8 +392,7 @@ class DsInstance(service.Service):
 
     def _get_replication_manager(self):
         # Always connect to self over ldapi
-        ldap_uri = ipaldap.get_ldap_uri(protocol='ldapi', realm=self.realm)
-        conn = ipaldap.LDAPClient(ldap_uri)
+        conn = ipaldap.LDAPClient.from_realm(self.realm)
         conn.external_bind()
         repl = replication.ReplicationManager(
             self.realm, self.fqdn, self.dm_password, conn=conn
@@ -557,45 +526,71 @@ class DsInstance(service.Service):
         )
 
     def __create_instance(self):
-        pent = pwd.getpwnam(DS_USER)
-
         self.backup_state("serverid", self.serverid)
 
-        self.sub_dict['BASEDC'] = self.realm.split('.')[0].lower()
-        base_txt = ipautil.template_str(BASE_TEMPLATE, self.sub_dict)
-        logger.debug("%s", base_txt)
+        # The new installer is api driven. We can pass it a log function
+        # and it will use it. Because of this, we can pass verbose true,
+        # and allow our logger to control the display based on level.
+        sds = SetupDs(verbose=True, dryrun=False, log=logger)
 
-        target_fname = paths.DIRSRV_BOOT_LDIF
-        base_fd = open(target_fname, "w")
-        base_fd.write(base_txt)
-        base_fd.close()
+        # General environmental options.
+        general_options = General2Base(logger)
+        general_options.set('strict_host_checking', False)
+        # Check that our requested configuration is actually valid ...
+        general_options.verify()
+        general = general_options.collect()
 
-        # Must be readable for dirsrv
-        os.chmod(target_fname, 0o440)
-        os.chown(target_fname, pent.pw_uid, pent.pw_gid)
+        # Slapd options, ie instance name.
+        slapd_options = Slapd2Base(logger)
+        slapd_options.set('instance_name', self.serverid)
+        slapd_options.set('root_password', self.dm_password)
+        slapd_options.verify()
+        slapd = slapd_options.collect()
 
-        inf_txt = ipautil.template_str(INF_TEMPLATE, self.sub_dict)
-        logger.debug("writing inf template")
-        inf_fd = ipautil.write_tmp_file(inf_txt)
-        inf_txt = re.sub(r"RootDNPwd=.*\n", "", inf_txt)
-        logger.debug("%s", inf_txt)
-        args = [
-            paths.SETUP_DS_PL, "--silent",
-            "--logfile", "-",
-            "-f", inf_fd.name,
-        ]
-        logger.debug("calling setup-ds.pl")
+        # Create userroot. Note that the new install does NOT
+        # create sample entries, so this is *empty*.
+        userroot = {
+            'cn': 'userRoot',
+            'nsslapd-suffix': self.suffix.ldap_text()
+        }
+
+        backends = [userroot]
+
+        sds.create_from_args(general, slapd, backends, None)
+
+        # Now create the new domain root object in the format that IPA expects.
+        # Get the instance ....
+
+        inst = DirSrv(verbose=True, external_log=logger)
+        inst.local_simple_allocate(
+            serverid=self.serverid,
+            ldapuri=ipaldap.get_ldap_uri(realm=self.realm, protocol='ldapi'),
+            password=self.dm_password
+        )
+
+        # local_simple_allocate() configures LDAPI but doesn't set up the
+        # DirSrv object to use LDAPI. Modify the DirSrv() object to use
+        # LDAPI with password bind. autobind is not available, yet.
+        inst.ldapi_enabled = 'on'
+        inst.ldapi_socket = paths.SLAPD_INSTANCE_SOCKET_TEMPLATE % (
+            self.serverid
+        )
+        inst.ldapi_autobind = 'off'
+
+        # This actually opens the conn and binds.
+        inst.open()
+
         try:
-            ipautil.run(args)
-            logger.debug("completed creating DS instance")
-        except ipautil.CalledProcessError as e:
-            raise RuntimeError("failed to create DS instance %s" % e)
+            ipadomain = IpaDomain(inst, dn=self.suffix.ldap_text())
+            ipadomain.create(properties={
+                'dc': self.realm.split('.')[0].lower(),
+                'info': 'IPA V2.0',
+            })
+        finally:
+            inst.close()
 
-        # check for open port 389 from now on
-        self.open_ports.append(389)
-
-        inf_fd.close()
-        os.remove(paths.DIRSRV_BOOT_LDIF)
+        # Done!
+        logger.debug("completed creating DS instance")
 
     def __update_dse_ldif(self):
         """
@@ -699,7 +694,6 @@ class DsInstance(service.Service):
         self._ldap_mod("memberof-conf.ldif")
 
     def init_memberof(self):
-
         if not self.run_init_memberof:
             return
 
@@ -708,15 +702,9 @@ class DsInstance(service.Service):
         dn = DN(('cn', 'IPA install %s' % self.sub_dict["TIME"]), ('cn', 'memberof task'),
                 ('cn', 'tasks'), ('cn', 'config'))
         logger.debug("Waiting for memberof task to complete.")
-        ldap_uri = ipaldap.get_ldap_uri(self.fqdn)
-        conn = ipaldap.LDAPClient(ldap_uri)
-        if self.dm_password:
-            conn.simple_bind(bind_dn=ipaldap.DIRMAN_DN,
-                             bind_password=self.dm_password)
-        else:
-            conn.gssapi_bind()
-        replication.wait_for_task(conn, dn)
-        conn.unbind()
+        with ipaldap.LDAPClient.from_realm(self.realm) as conn:
+            conn.external_bind()
+            replication.wait_for_task(conn, dn)
 
     def apply_updates(self):
         schema_files = get_all_external_schema_files(paths.EXTERNAL_SCHEMA_DIR)
@@ -838,6 +826,8 @@ class DsInstance(service.Service):
             dsdb.create_from_pkcs12(self.pkcs12_info[0], self.pkcs12_info[1],
                                     ca_file=self.ca_file,
                                     trust_flags=trust_flags)
+            # rewrite the pin file with current password
+            dsdb.create_pin_file()
             server_certs = dsdb.find_server_certs()
             if len(server_certs) == 0:
                 raise RuntimeError("Could not find a suitable server cert in import in %s" % self.pkcs12_info[0])
@@ -854,6 +844,8 @@ class DsInstance(service.Service):
             self.add_cert_to_service()
         else:
             dsdb.create_from_cacert()
+            # rewrite the pin file with current password
+            dsdb.create_pin_file()
             if self.master_fqdn is None:
                 ca_args = [
                     paths.CERTMONGER_DOGTAG_SUBMIT,
@@ -880,7 +872,7 @@ class DsInstance(service.Service):
                     profile=dogtag.DEFAULT_PROFILE,
                     dns=[self.fqdn],
                     post_command=cmd,
-                    resubmit_timeout=api.env.replication_wait_timeout
+                    resubmit_timeout=api.env.certmonger_wait_timeout
                 )
             finally:
                 if prev_helper is not None:
@@ -895,22 +887,33 @@ class DsInstance(service.Service):
             if prev_helper is not None:
                 self.add_cert_to_service()
 
-        dsdb.create_pin_file()
-
         self.cacert_name = dsdb.cacert_name
 
-        ldap_uri = ipaldap.get_ldap_uri(self.fqdn)
-        conn = ipaldap.LDAPClient(ldap_uri)
-        conn.simple_bind(bind_dn=ipaldap.DIRMAN_DN,
-                         bind_password=self.dm_password)
+        # use LDAPI?
+        conn = ipaldap.LDAPClient.from_realm(self.realm)
+        conn.external_bind()
 
-        mod = [(ldap.MOD_REPLACE, "nsSSLClientAuth", "allowed"),
-               (ldap.MOD_REPLACE, "nsSSL3Ciphers", "default"),
-               (ldap.MOD_REPLACE, "allowWeakCipher", "off")]
-        conn.modify_s(DN(('cn', 'encryption'), ('cn', 'config')), mod)
+        encrypt_entry = conn.make_entry(
+            DN(('cn', 'encryption'), ('cn', 'config')),
+            nsSSLClientAuth=b'allowed',
+            nsSSL3Ciphers=b'default',
+            allowWeakCipher=b'off'
+        )
+        try:
+            conn.update_entry(encrypt_entry)
+        except errors.EmptyModlist:
+            logger.debug(
+                "cn=encryption,cn=config is already properly configured")
 
-        mod = [(ldap.MOD_ADD, "nsslapd-security", "on")]
-        conn.modify_s(DN(('cn', 'config')), mod)
+        conf_entry = conn.make_entry(
+            DN(('cn', 'config')),
+            # one does not simply uses '-' in variable name
+            **{'nsslapd-security': b'on'}
+        )
+        try:
+            conn.update_entry(conf_entry)
+        except errors.EmptyModlist:
+            logger.debug("nsslapd-security is already on")
 
         entry = conn.make_entry(
             DN(('cn', 'RSA'), ('cn', 'encryption'), ('cn', 'config')),
@@ -941,10 +944,8 @@ class DsInstance(service.Service):
                             subject_base=self.subject_base)
         trust_flags = dict(reversed(dsdb.list_certs()))
 
-        ldap_uri = ipaldap.get_ldap_uri(self.fqdn)
-        conn = ipaldap.LDAPClient(ldap_uri)
-        conn.simple_bind(bind_dn=ipaldap.DIRMAN_DN,
-                         bind_password=self.dm_password)
+        conn = ipaldap.LDAPClient.from_realm(self.realm)
+        conn.external_bind()
 
         nicknames = dsdb.find_root_cert(self.cacert_name)[:-1]
         for nickname in nicknames:
@@ -975,14 +976,9 @@ class DsInstance(service.Service):
         dsdb = certs.CertDB(self.realm, nssdir=dirname,
                             subject_base=self.subject_base)
 
-        ldap_uri = ipaldap.get_ldap_uri(self.fqdn)
-        conn = ipaldap.LDAPClient(ldap_uri)
-        conn.simple_bind(bind_dn=ipaldap.DIRMAN_DN,
-                         bind_password=self.dm_password)
-
-        self.export_ca_certs_nssdb(dsdb, self.ca_is_configured, conn)
-
-        conn.unbind()
+        with ipaldap.LDAPClient.from_realm(self.realm) as conn:
+            conn.external_bind()
+            self.export_ca_certs_nssdb(dsdb, self.ca_is_configured, conn)
 
     def __add_default_layout(self):
         self._ldap_mod("bootstrap-template.ldif", self.sub_dict)
@@ -1034,11 +1030,6 @@ class DsInstance(service.Service):
             str(self.subject_base)
         )
 
-    def __enable_ldapi(self):
-        self._ldap_mod("ldapi.ldif", self.sub_dict,
-                       ldap_uri="ldap://localhost",
-                       dm_password=self.dm_password)
-
     def __enable_sasl_mapping_fallback(self):
         self._ldap_mod("sasl-mapping-fallback.ldif", self.sub_dict)
 
@@ -1081,16 +1072,22 @@ class DsInstance(service.Service):
 
         try:
             self.fstore.restore_file(paths.LIMITS_CONF)
+        except ValueError as error:
+            logger.debug("%s: %s", paths.LIMITS_CONF, error)
+
+        try:
             self.fstore.restore_file(paths.SYSCONFIG_DIRSRV)
         except ValueError as error:
-            logger.debug("%s", error)
+            logger.debug("%s: %s", paths.SYSCONFIG_DIRSRV, error)
 
         # disabled during IPA installation
         if enabled:
+            logger.debug("Re-enabling instance of Directory Server")
             self.enable()
 
         serverid = self.restore_state("serverid")
         if serverid is not None:
+            # What if this fails? Then what?
             self.stop_tracking_certificates(serverid)
             logger.debug("Removing DS instance %s", serverid)
             try:
@@ -1099,19 +1096,24 @@ class DsInstance(service.Service):
                 logger.error("Failed to remove DS instance. You may "
                              "need to remove instance data manually")
 
-            installutils.remove_keytab(paths.DS_KEYTAB)
-            installutils.remove_ccache(run_as=DS_USER)
+        else:
+            logger.error("Failed to remove DS instance. No serverid present "
+                         "in sysrestore file.")
 
+        ipautil.remove_keytab(paths.DS_KEYTAB)
+        ipautil.remove_ccache(run_as=DS_USER)
+
+        if serverid is None:
             # Remove scripts dir
             scripts = paths.VAR_LIB_DIRSRV_INSTANCE_SCRIPTS_TEMPLATE % (
                 serverid)
-            installutils.rmtree(scripts)
+            ipautil.rmtree(scripts)
 
             # remove systemd unit file
             unitfile = paths.SLAPD_INSTANCE_SYSTEMD_IPA_ENV_TEMPLATE % (
                 serverid
             )
-            installutils.remove_file(unitfile)
+            ipautil.remove_file(unitfile)
             try:
                 os.rmdir(os.path.dirname(unitfile))
             except OSError:
@@ -1183,9 +1185,12 @@ class DsInstance(service.Service):
         dirname = config_dirname(serverid)[:-1]
         dsdb = certs.CertDB(self.realm, nssdir=dirname)
         if dsdb.is_ipa_issued_cert(api, nickname):
-            dsdb.track_server_cert(nickname, self.principal,
-                                   dsdb.passwd_fname,
-                                   'restart_dirsrv %s' % serverid)
+            dsdb.track_server_cert(
+                nickname,
+                self.principal,
+                password_file=dsdb.passwd_fname,
+                command='restart_dirsrv %s' % serverid,
+                profile=dogtag.DEFAULT_PROFILE)
         else:
             logger.debug("Will not track DS server certificate %s as it is "
                          "not issued by IPA", nickname)
@@ -1215,7 +1220,8 @@ class DsInstance(service.Service):
         # shutdown the server
         self.stop()
 
-        dirname = config_dirname(installutils.realm_to_serverid(self.realm))
+        dirname = config_dirname(
+            ipaldap.realm_to_serverid(self.realm))
         certdb = certs.CertDB(
             self.realm,
             nssdir=dirname,
@@ -1239,9 +1245,12 @@ class DsInstance(service.Service):
         return status
 
     def __root_autobind(self):
-        self._ldap_mod("root-autobind.ldif",
-                       ldap_uri="ldap://localhost",
-                       dm_password=self.dm_password)
+        self._ldap_mod(
+            "root-autobind.ldif",
+            ldap_uri=ipaldap.get_ldap_uri(realm=self.realm, protocol='ldapi'),
+            # must simple bind until auto bind is configured
+            dm_password=self.dm_password
+        )
 
     def __add_sudo_binduser(self):
         self._ldap_mod("sudobind.ldif", self.sub_dict)
@@ -1352,7 +1361,7 @@ class DsInstance(service.Service):
 
 def write_certmap_conf(realm, ca_subject):
     """(Re)write certmap.conf with given CA subject DN."""
-    serverid = installutils.realm_to_serverid(realm)
+    serverid = ipaldap.realm_to_serverid(realm)
     ds_dirname = config_dirname(serverid)
     certmap_filename = os.path.join(ds_dirname, "certmap.conf")
     shutil.copyfile(
